requirements.txt
flask
numpy
pandas
opencv-python
Pillow
tensorflow
deepface
tf-keras
scikit-learn